server_code/requirements.txt
airtable-python-wrapper==0.15.3
opencv-python
pyzbar